refactor(frontend): split cd stream hardware
